Replace user visible usage of "pinned" with "starred".
Also adds the gradient photos because I am a goon..
