+++ pages/dao/[contractAddress]/proposals/[proposalId].tsx
@@ -20,7 +20,7 @@ const Proposal: NextPage = () => {
       <div className="w-full col-span-4 p-6">
         <Breadcrumbs
           crumbs={[
-            ['/dao/list', 'DAOs'],
+            ['/starred', 'Home'],
             [`/dao/${contractAddress}`, sigInfo.config.name],
             [router.asPath, `Proposal ${proposalId}`],
           ]}
