+++ pages/multisig/[contractAddress]/index.tsx
@@ -93,7 +93,7 @@ const Home: NextPage = () => {
           <div className="flex justify-between items-center">
             <Breadcrumbs
               crumbs={[
-                ['/pinned', 'Home'],
+                ['/starred', 'Home'],
                 [router.asPath, sigInfo.config.name],
               ]}
             />
