+++ pages/multisig/[contractAddress]/proposals/create.tsx
@@ -76,7 +76,7 @@ const ProposalCreate: NextPage = () => {
       <div className="w-full col-span-4 p-6">
         <Breadcrumbs
           crumbs={[
-            ['/multisig/list', 'Multisigs'],
+            ['/starred', 'Home'],
             [`/multisig/${contractAddress}`, sigInfo.config.name],
             [router.asPath, 'Create proposal'],
           ]}
